RequirementMachine: Move RequirementMachine.h to lib/AST/RequirementMachine
